剪贴板内存优化 Signed-off-by:yang_yong_qi<yangyongqi4@h-partners.com>

+++ services/dfx/src/calculate_time_consuming.h
@@ -23,7 +23,7 @@
 
 namespace OHOS {
 namespace MiscServices {
-enum class TimeLevel : std::uint64_t {
+enum class TimeLevel : std::uint32_t {
     PER_FIVE_HUNDRED_MS_ZERO = 0,
     PER_FIVE_HUNDRED_MS_ONE,
     PER_FIVE_HUNDRED_MS_TWO,
